Banished the possibility of Show Pos
You are right, that instance does not belong there. Show and Read are
for converting to and from strings that you can paste into code to
recover the same value. Pretty-printing is to be done in other ways. (I
realize you probably only didnt do it via Show and Read because (Int,
Int) already has an instance for that :P)