Tighten another recursion, notice FAIL
evalPredictive takes an evaluator and lifts it by a one-turn prediction. This is semantically identical to what you did.

Now for the FAIL: You presumably are trying to implement Minmax. If you've never heard that word, and are inventing a way to automate chessplaying from scratch, and want to think of the interesting parts yourself, don't read on.

You predict the game as if both players are playing maximally in your favor, cooperating to get the board is good in the value of the AI side's evaluator as possible. Minmax says that the other player is trying to maximize his own score, and so you should run the evaluator as if the AI was playing the other side, meaning that you should replace the last s in 164 by (other s).